refactor(assets): extract shared sdk-lib.mjs for ensure-sdk / generate-sdk
Both scripts duplicated the dist-bundling fallback and (for generate-sdk) the
spec-hash comparison. Move that logic into sdk-lib.mjs so each script is a
thin wrapper; EE reuses this same file by relative import instead of
duplicating it across repos.
